Invert aliases map
The former solution kinda worked, but failed for one crucial use-case,
when we need to call the same GraphQL method twice for different
aliases. This is the main use-case that makes sense though. Renaming
fields within Domain types is the odd one out with questionable/rare
actual use.

With the inversion, the main use case works. This code introduces a new
potential issue though, which is that we basically map the aliases as
globals for the request and so if you would repeat the same alias deeper
inside the query things will go wrong.

So this only works if the used aliases are unique across the whole
GraphQL document. Given the aforementioned weakness of those deep alias
use cases, this seems like a reasonable compromise. And is also a
compromise that removes the need for a large-scale rewrite with a more
complicated data structure.

Adjusts the tests to use the new convention.