Revalidate scalar values
At first I left this out intentionally, since I was thinking of GraphQL
typed data and field types that don't change.

But we do support editing just observables and we do use that to edit
our condition graphs and those graphs or rather the generic scalar
inside them do change. And this would happen for GraphQL typed data with
embedded GenericScalar, too.

Since we value we get back is the value as-is, we can't just validate
what could be a custom scalar value.

But we do have the string equivalence etc, so we can convert the value
to string first and then validate it and reasonably expect that not to
work with a wrong value.