Refactor Field onChange
Extend the signature of the onChange handler from (fieldContext, value) =>
... to ({oldValue,fieldContext}, value) => ..., moving the field context
to the property "fieldContext" in a new context object and add the old
value.

This is not to expensive anyway and we only need to do so if a field
change handler is defined.

Harmonizes the onChange handler in <Select/>. I think the prevention was
a bad idea and bad UI. Use a separate observable field if you want
to do similar stuff.

Corrects propTypes to reflect available props (spread to <Field/>)