Extend helpText
Enable helpText to be react elements or function that renders content
and make sure that the helpText captured in the field context gets
updated correctly.

Note that this requires a *stable helpText reference*. If it is more
than a string, the reference to the function or react element needs to
be stable to prevent unescessary field context recreation.

Adds test case