Refactor handling of intermediary objects
For flat objects, we disable the whole form when the form root is null.
It follows that it makes sense to disable fields on intermediary objects
that are missing.

In some ways, this already worked in that the undefined value the code
reads for a missing object is converted into a null/empty value.

Switches a NORMAL field to DISABLED when updating the component while
the parent object is missing.

Ignores fields if their mode is not NORMAL. This should already have
been done, now it includes the missing parent object.

Adds two tests for missing intermediary objects.