Refactor validation
So far we've been invoking the local <Field/> validation and the
validateField method of the high-level validation with the String input
provided by the user, after validating it successfully.

For a lot of fields this makes no difference, but for e.g. Calendar
fields it seems silly that the validator should have to redo the
conversion from string.

The recent scalar revalidation also introduced an error here, already
thinking we used scalar values there.

This cleans the whole thing up by converting the string value sooner so
that the converted value can be passed on to the local and hl
validation.

Moves the scalar revalidation into FormContext.revalidate which is a
better location since we only need the revalidation in that case because
coming from handleChange we already just finished the scalar validation
successfully.

Changes prop-type documentation.

Adds an actual test-case for the scalar revalidation catching dynamic
type changes.

Changes the corresponding pre-existing tests to our current
expectations, mostly that "" is now passed as scalar value `null`