Fix boolean editing
A detail that escaped my attention was that we still had some special
case handling for booleans which we inherited from the time when we used
formik and which largely remained that way for the same reasons with the
new code.

The way to edit booleans is a checkbox and that checkbox wants true
booleans as its special checked prop and it reports those changes as
true booleans to handleChange. Which is why it is more convenient to
keep the scalar value we use on the JavaScript side to be a true
boolean.

Which is not a string like the new scalar revalidation feature assumed.

This fixes that.

This bug could survive because the test, which did exist, actually did
not check that there were no errors after resubmitting the changed
checkbox. (It also kept isolating the form object and not rechecking
that)

This fixes both deficits.