FormContext: Make revalidate optionally async
This one is a bit weird. So we have async validation that is triggered
alongside our normal validation, but which so far never gets actually
waited on from the caller side.

However to correctly regard the complete validation status we do need to
wait until all the async validations have returned.

This commit tries to make the best of this awkward situation. We keep
the sync revalidation sync and then return a Promise that resolves when
all the async validations we have triggered have concluded. If any of
them reject, the overall Promise is rejected, too, but it always waits
until all of them are concluded.

For our purposes, this corresponds with a change in automaton's
<Button/> component which now waits for the async validation.

We add an additional method `waitForAsyncValidation` which waits for
the current overall async validation promise if such a promise is still
active. This is useful to prevent the awkwardness of being sure we just
triggered a revalidation with the form submit, but not wanting to
trigger another revalidation just to wait for it.

This way we can just wait for the one we are sure we have triggered to
conclude. If we were wrong about this, the method returns a simple
`Promise.resolve()`