Quick-Fix Process change/Errors
So we have the awkward situation that when we leave a process that is
currently in an error state to enter another process or the same
process with different parameters, the following happened:

We started a new root process, we unregistered all forms, but the error
which was associated with the form root object actually remained in the
FormContext storage. Technically this is not even wrong since the old
process is not 100% dead, as the user can use browser navigation to go
back.

However we are now in the situation where our usually global form
context is poisoned with error messages which prevent us from working
with our forms without it being any kind of visible *where* the error is
(it is in the inactive process), only that there is an error.

To fix or evade this problems we do two things:

After the user has given us permission to destroy the current dirty
state, we actually do so by calling workingSet.undo() if such a working
set exist.

And we remove all errors in the global form context, which is not 100%
kosher but necessary. If the undoing of the workingSet did not fix all
problems in the inactive process, another revalidate() should reveal
them again. Or otherwise the process should be designed so that that is
the case.