GlobalErros now shows all errors
Until now only the errors of the root form were shown, but the
requirement states to show all errors from the currently loaded forms.
Also the element resolver was fixed to actually search the document, not
only the forms containing the GlobalErrors, since the GlobalErrors can
be in its own form never able to access the forms the errornous elements
are residing in. This is to fix the focus behaviour when clicking an
error entry.